requirements.txt
ragas~=0.2.0
datasets~=2.20
openai~=1.51
anthropic~=0.34
python-pptx~=1.0
PyYAML>=6.0.1,<7
